Refactoring SimpleUncheckedException
