+++ src/main/java/io/github/sashirestela/openai/exception/SimpleOpenAIException.java
@@ -1,15 +1,15 @@
-package io.github.sashirestela.openai;
+package io.github.sashirestela.openai.exception;
 
 import java.text.MessageFormat;
 import java.util.Arrays;
 
-public class SimpleUncheckedException extends RuntimeException {
+public class SimpleOpenAIException extends RuntimeException {
 
-    public SimpleUncheckedException(String message) {
+    public SimpleOpenAIException(String message) {
         super(message);
     }
 
-    public SimpleUncheckedException(String message, Object... parameters) {
+    public SimpleOpenAIException(String message, Object... parameters) {
         super(MessageFormat.format(message, Arrays.copyOfRange(parameters, 0, parameters.length - 1)),
                 (Throwable) parameters[parameters.length - 1]);
     }
